+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_ca.dart
@@ -556,10 +556,17 @@ class StreamChatLocalizationsCa extends GlobalStreamChatLocalizations {
   String get updateYourCommentLabel => 'Actualitzar el vostre comentari';
 
   @override
-  String get enterYourCommentLabel => 'Introduïu el vostre comentari';
+  String get enterYourCommentLabel => 'Introdueix el teu comentari';
+
+  @override
+  String get endVoteConfirmationText =>
+      'Estàs segur que vols finalitzar la votació?';
 
   @override
   String get createLabel => 'Crear';
+
+  @override
+  String get endLabel => 'Finalitzar';
 
   @override
   String pollVotingModeLabel(PollVotingMode votingMode) {